Add a check for unmatching mapping and coords file
The script interface will now display an error when there are no samples
in common between the mapping file and the coords file.

The script interface will now display an error when the samples
represented in the coordinates file are not all represented by the
mapping file. The user however can override this with an extra flag.

This should fix issue #21