Add jackknifing option to the script interface
The script interface is now capable of producing jackknifed plots, an
option to select what method you want to use for this calculation was
added as well.

The main input option, no longer takes a single file only, now it is
capable of taking a directory with multiple files in it. Several
different checks are being performed from the script interface.

This should fix #11