Fix bug in preprocess_otu_table
This bug caused the biplot information to be placed in the exact
coordinates where samples were positioned. This function now makes use
of emperor.sort to make sure that the sample identifiers match between
coordinates files and otu tables.

The script interface was updated accordingly since the filtering of the
otu table is now performed using the header in the coordinates data.