Fix bug reported in #66
The copy_support_files function was not accounting for filepaths that
contain spaces or other invalid characters, hence when trying to run
make_emperor.py and setting the output path to something like
"/Users/jappleseed/data/oct 12/" would fail because of the "oct 12"
component in the path.

A new test has been added to test_util.py.

Fix #66