Change the values of g_min and g_max for all axes
These values now have an extra padding that's relative to the size of
the spheres (6 times the sphere radius), this is with the intent of
making the axes in the plot look a little bit better. This change also
avoids cases when a sphere would be crossed by an axis line.

Test suite was updated accordingly.

Fixes #58