Fix with --ignore_missing_samples & --custom_axes
When passing --ignore_missing_samples and trying to add a custom axis,
the lack of samples in the coords data would make the data in the
mapping file not match the dimensions of the data in the PCoA data.

Change the default value for --add_vectors and change the contition for
the formatting of the footer.

Change __copyright__ string