Add new identifying way for elements
HTML elements were identified using the value in the mapping file for
the needed column, this caused that when there was a value that had
disallowed characters, the coloring would be misleading and wrong.

This was true for both label coloring and sample coloring, for further
discussion see #99

Fixes #99 and fixes #15.