Add new version of Three.js
Current release we are using is the one tagged in the repository as r58,
doing git checkout r58 in that repo should suffice to retrieve this
version of the files.

Through the many updates they made, that affect emperor.js:
    - Removed TrackbalControls and from the main build, now lives in
      the examples folder.
    - Deprecated Vertex in favor of Vector3.
    - Removed the method setHSV.
    - Changed the method getHEX to return a hexadecimal value and not
      the hexadecimal string that we were expecting, should now use
      getHEXString.
    - Added a new class called ColorConverter that should be used to
      convert between HSV -> RGB and HSV -> HSL etc.

Just as a way to document this, the version in which we were was r47
however the build we had didn't match with the build that was in the
repo for that release, so I imagine it was really r47-dev.