Change emperor.js to support THREE.js r58
The color for the renderer has to be set now as a THREE.Color and not as
a hexadecimal string, hence requiring a change for the color picker in
the background and in the init function.

Vertex objects were deprecated; so makeLine() had to be modified so it
now uses only Vector3 objects.

Matrix multiplications were removed from the Matrix4 objects,
multiplications are now done using Vector3.Matrix4 or
Vector3.applyProjection() when the factor to multiply for is a vector or
Matrix4.multiplyMatrices() to multiply two matrices; these changes are
refelected in toScreenXY();

Colors no longer return a string in hexadecimal format when calling
getHex, hence all THREE.Color objects now call the method
getHexString().

Additionaly the credit strings at the top were updated.

Fixes #84.