Add ability to process input data for comparisons
preprocess_coords_data adds a new functionality to deal with the case
where the input is a set of coordinate files and the intended use is for
a comparison plot. This function now also return the number of of the
mapping file that will be needed.

preprocess_mapping file adds the ability to clone the data in the
mapping file as many times as requested by passing a new parameter; this
in turn will add a suffix to each replication.

Test suite was also updated for the util.py module