Change URLs to point to emperor.colorado.edu
The usage examples are now hosted in emperor.colorado.edu, all examples
should be accessible from the README.md file of this branch and once
merged from the README.md of the main repository.

Note that the last of the examples is broken because of a bug with the
parallel plots.

Fixes #96