Add extra ordering step for plot comparisons
When the plot comparison is one with multiple procrustes transformations
the code now guesses the master file and forces it to be the first in
the list of samples when displayed in the interface.

This was already the case but with the added lines this is _going_ to be
the case always.

-------------------------------------------------------------------------

Remove typos and empty lines.