Fix bug with comparison plots
Specifying the reference set of coordinates was previously taking the
effect of making a random set of coordinates the reference instead of
the one passed from the CLI. This has been corrected.

Change the way filenames were being sorted for comparison plots, these
are now sorted based on the suffix they have if they are in a
QIIME-compliant format.

Thanks @nbokulich for pointing thes out!