Fix bug where custom axes & comparisons would fail
The coordinates appended a suffix to each of the sample identifiers
depending of which replicate it was (i. e. for the first set of
coordinates the appended string will be '_0', for the second set it
would be '_1', etc ...) processing, hence when trying to match the
sample identifiers to the ones found in the mapping file, it would fail
as there would be no exact matches. A mismatching of this nature would
cause an exception with the following message:
    ValueError: concatenation of zero-length sequences is impossible

Also fixed a typo in a command.