Add animation menu+elements and formatting
These changesets add an animation tab in the main GUI that has the
bare-bone controllers as ported from E-vident SVN revision 253.

The formatter of the mapping file now selects which of the available
columns are numeric, if a column is numeric then it is saved else it is
not saved.