Add message about Google Chrome being 32-bit
I've added a message in red to the documentation's index explaining this
problem (for more details, see below).

After an e-mail exchange with Chromium engineers we've figured that
there are some cases where the amount of samples in a study requires
a 64-bit version of Google Chrome which is only available when running
under Linux as that's the only officially supported platform for 64-bit
builds of Google Chrome. See the following link:
    https://code.google.com/p/chromium/issues/detail?id=312958

There are open tickets in their issue tracker trying to target the
release of a 64-bit version for OS X, however this has not been
addressed yet:
    https://code.google.com/p/chromium/issues/detail?id=18323

It's worth adding that in previous versions of Google Chrome this didn't
matter and visualizations would work without a problem, hence the
suggestion now is to open the bigger studies using Safari or Firefox
(which we've tested and work just fine).