Change styling of the tables to be equal
In the "Scaling" and the "Visibility" tab, the handle for the sliders
was making the table show sliders when it reached the far left or the
far right. This was corrected by adding some padding to each of the
sides, hence now nothing needs to be resized.

The "master" scaling and transparency sliders (the ones at the bottom of
each of these tabs) were showing bigger as compared to the other sliders
(once the padding was added); adding an extra padding to the cells that
contain these sliders resolve this inconsistency. Note that since this
took place in format.py the tests had to be updated.