Change all_tests.py to support qiime-deploy
Given the changes that were introduced in this branch, there's no longer
a need to deploy a full QIIME installation and hence we created
Emperor's own configuration file so that Jenkins can run the full test
suite. This in turn broke the way the script_usage_tests were being
executed.