Add checks for correct formatting of pcoa files
Although there were a couple checks in place, the parser was not
verifying the format and loading the data in a fairly naive fashion.
This is no longer the case and will show the user an error.

Additionally I moved a block of code that was being executed immediatly
after the step that could raise an exception.

Fixes #216