Remove fog from the main scene in the 3d renderer
When the scene was being initialized, the GL_FOG property was enabled,
which would make spaces of around 1000 units in width per dimension to
be lost in the fog given that the density was rather high for such
spaces. This did not represent a problem for cases where the data being
ploted was as generated by distance matrices where the values were small
because the generated spaces would never exceed a couple units per
dimension.

In the code GL_FOG was enabled as GL_EXP_2, from the GL documentation:

http://www.opengl.org/sdk/docs/man2/xhtml/glFog.xml
------------------------------------------------------------------------
Fog blends a fog color with each rasterized pixel fragment's
post-texturing color using a blending factor `f`. Factor `f` is computed
in one of three ways, depending of the fog mode.

[...]

The equation for GL_EXP2 fog is:

f = e^(-(density*c))^2
------------------------------------------------------------------------

I decided to completely disable the fog effect from the main scene as it
is irrelevant for a three dimensional plot and it's mostly useful for
cases where you want to give the effect of light faiding away as
distance increases. The alternative to this decision was to take the
matrix inputed in the commmand line interface and scale it to a proper
value, but this would increase the processing time and the fog doesn't
really add any value (qualitative) to the plots.

Fixes #220