Add changes to all_tests.py to run JS tests
All tests is now capable of running the JavaScript test suite and report
the failures on the command line. This script depends on phantomjs to
run the tests and assumes that this program is available on the PATH of
the current session.

Additionally runner.js was also added to respond to the QUnit callbacks
and appropriately format the failure, success, etc messages.