BUG/ENH: Fix broken CSS property for media buttons
During the huge CSS and stylesheet refactoring of 2014 when all the
classes were cleaned up and emperor.css became a lot more organized, I
also added one space between a class name and an event, hence the media
buttons used for the animations would not change their background color
on hover. Removing this space brings back the old functionality and
makes it look a lot more natural.

Moral of the story: One space can make a huge difference and it may take
you a couple months to realize it was all your fault in the first place.