ENH/TST: Add category name to TrajectoryOfSamples
In order for the trajectories to be correctly colored, we need to know
which category were they formed from, therefore an easy way to do this
is to take the include this information as part of the metadata of the
trajectory itself.

The test suite has been updated accordingly to accomodate these changes.