ENH: Use different colormaps for coloring
Using chroma we can use colorbrewer colormaps for gradient coloring.
I've moved the choose-color function into the util.js module and more
importantly I have removed the dependency on THREE.Color class. We were
loosely using this class and converting back and forth everywhere, and
we didn't need to do this.

The licenses for colorbrewer and for chroma were added to the LICENSE.md
file.