ENH: Bring back the QIIME colors - fixes #310
Discrete coloring using the QIIME-defined colors is now an option in the
colormap selector under the colors tab.

getDiscreteColor now takes an optional "map" argument to know whether it
is using the ColorBrewer or the QIIME list of colors, defaults to
ColorBrewer as that is Emperor's default.

Finally there's a note in the ChangeLog describing this change.

Fixes #310.