ENH: Change _bodyGrid to bodyGrid
This property shouldn't be private as the usage of it is more natural
for subclasses through direct access to the grid.

Additionally, this patch fixes a few places where the scope of `this`
was being used incorrectly and buildGrid no longer returns a grid, but
instead initializes the grid object to function accordingly.