MAINT: Overall improvements to EmperorController
Fix the resize method of the EmperorController object to properly fit
contents to the size of the screen, these had been a long-time source of
headaches for the past version of Emperor. Turns out we just had to
account of the size of the tabs list element.

Change the way we refer to a few DOM elements, we now refer to them
through property names of the class, namely the following changes
happened:

- plotSpaceId -> plotSpace: the Id suffix was incorrect as in reality it
  is just a DOM object.
- Create _$tabsContainer and _$tabsList to replace two divs that were
  previously being accessed with the following ids:
  'emperor-tabs-container' and 'emperor-tabs-list'. I've left the name
  so that we can distinguish them when we debug the application, but
  they should be only accessed through the property itself.

Finally, the docstrings for some methods were formatted incorrectly so
this was changed to fit our standard.