TST/BUG: Fix resizing, alignment and tests
Resizing: The size of the body now will change appropriately as needed
when the container is resized, more importantly this will change and
update the size of the slick grid of the EmperorAttributeViewController
class. I've moved this logic to the EmperorViewControllerABC and called
the super method from the subclass.

Alignment: the body and header divs will now center themselves in their
main container.

Tests: Our tests for resize, toJSON and fromJSON were giving false
positives, I've now fixed this and added the appropriate tests for the
resize method in the base class.