BUG: Fix problem with require.js baseUrl attribute
As it turns out, when you set this configuration attribute for
require.js, you are setting a global attribute, not a local one, which
is fine if you are the only application using require.js to load
resources. However this becomes problematic when you have multiple
applications trying to do this and setting the baseUrl. The end result
is a race where whatever application got to load first is the one that
will impose the base_Url.

We've removed the attribute altogether and replaced it for prefixes in
the path names. This is a bit clunkier in the sense that the jinja
template now has to write the information for each of the configuration
files.

HT to @tanaes who brought up that it made _no_ sense for a module to
"die", which later made me realize what the problem was.