ENH: Add the ability to load new decompositions
New decompositions can be loaded from the main controller, these update
the view controllers and the view controllers in return are capable of
adding new metadata categories in themselves.

There's still a few bugs, the main one originating from the fact that
when we update the metadata selection menu, we are re-building the
entire menu, instead of only inserting the needed keys, hence we'll need
to do something a bit more sophisticated.

Note also that the setMetadataField method in the
EmperorAttributeViewController class does not account for multiple
decompositions and will instead select the first decomposition view
where the category is found. This should instead be done on a view by
view basis, where the user can specify what the view it is that they
want to set the metadata category. Which leads to the problem of
selecting the appropriate value from the menu when this is needed. We
need to navigate to the appropriate element under the appropriate option
group to do this.