ENH: Checkboxes no longer require extra click
Previously, to hide/show sets of samples, you would need to click on the
visibility grid to "confirm" the change, this has been fixed to
auto-confirm. Since we were using the original minified editor file,
I've modified the code in-place, but we can re-add the un-minified
version if others think it's appropriate (no strong feelings).