ENH: Jackknifed Plots in new-api (#587)
* ENH: Factor out the processing of coordinates

* REF: Rework the underlying coordinate processing

Instead of doing everything from scratch and porting things slowly from
the previous version of emperor, for now, we will use the functions that
are available in util.py

* BUG: Fix bug with confidence intervals

There was a bug in the loop that formatted the data into the legacy
structures. It was copying the "master" ordination over and over.

* BUG: Fix small problem with pandas series

* API: API changes to DecompositionModel

It makes more sense to initialize the object with a dictionary of
objects instead of arguments.

* TST: Fix tests to use the new DecompositionModel

* ENH: ScenePlotView3D now deals with CIs

* ENH: Make ColorViewController aware of CIs

* ENH: VisibilityController is now aware of CIs

* ENH: Improve DecompositionView

* STY: Stylistic changes to the JavaScript code

* ENH: Confidence intervals can now be serialized

* ENH/TST: Add test for jackknifing

* BUG: Fix bad merge

* BUG/TST: Fix problem with settings string

* DBG: Experimental numpy modernization

* DBG: Travis pl0x

* TST: Improve tests

We only test two cases with exact formatting of the HTML/JavaScript,
most of the data processing is tested at the Python level, this resulted
in the addition of a get_template method.

* TST: Remove unused text and fix test name

* BUG: Fix bug in util.py

* BLD: Fix .travis.yml

* BUG/TST: User values because old pandas 🐼

* TST: add more tests

* DOC: Add Jackknifed example

* DOC: Add extra notes to all the notebooks

As per @antgonza's request.

* DOC: Remove plots from example notebooks

After debugging a problem with @antgonza, we noticed that the main
problem came from the fact that a pre-loaded plot would break the
environment for any future plots. This patch removes all saved plots and
adds a small description suggesting that users change remote=True if
they wish to share their plots via Github.