Update three.js and bug fixes (#602)
* ENH/BLD: Update to the latest version of THREE

* BUG: Fix bug with mixed color types

The code in emperor was mixing color representations, as hexadecimal
numbers and as strings. Which lead to the AxesViewController to stop
working when colors were saved as numbers and then tried to be loaded as
strings. This patch changes all properties to always deal with CSS-like
strings.

In addition, in the latest version of three.js the renderer's clearColor
property was deprecated and replaced by the scene's background property.

* BUG: Fix problems with SVGRenderer and opacity

* BUG: Fix problem with default color for scene