ENH: Make Emperor aware of biplots
This commit adds a fair number of changes, for the most part the intent
is to be able to interpret and process feature data as transported by
the ordination. The constructor can now take as a parameter feature
metadata. As a result when serializing the object into a JSON object, we
now use a single object, as returned by _to_dict. The main template was
also modified to account for this change. _to_dict puts the sample and
feature ordinations in a common format for easy consumption by the
JavaScript API.

A number of methods have been modified to become aware about feature
metadata and the feature ordination. Specifically:

- _validate_metadata: takes as a parameter the metadata it is validating
and allows for a case where no metadata is provided.

- _process_data: checks if there's a feature ordination and if there is
one proceeds to create the appropriate objects and convert them into
JSON-serializeable types.

- _to_legacy_map now takes as a parameter the metadata to convert.