ENH:Make the GUI objects aware of procrustes plots
Adds a couple of methods for the model and the views to process an edge
list that's used to display procrustes plots.

This patch also includes some cleanup to the repeated code in the
changeVisibleDimensions and flipDimension methods. See updatePositions.

Draw.js has a function to draw lines as a THREE.LineSegments object,
this is recommended from their documentation to be more efficient at
rendering, etc. This used to be a problem when we had a large number of
samples being compared at the same time.