BUG: Prevent input propagation in the notebook 📓 (#649)
The text inputs in the color picker propagated the keypress and keydown
events all the way down to the notebook. In the case where the inputs
conflicted with a shortcut in the notebook, the shortcut itself would be
triggered.

This has been fixed by preventing the propagation of events in all the
color pickers.