ENH: Fix HTML for axes and .ready for controller
The HTML code for the procrustes controls was missing a few closing
tags. I've also changed the formatting so it is easier to read.

The _expected and _seen attributes in the main controller are now
updated through the addTab method, so there's no need to update these
values manually.

Addressing @antgonza's review comments