Procrustes new api (#647)
* ENH: Add makeArrow function

* ENH: Add the type attribute to the decomp object

* ENH:Make ScenePlotView3D and DecompView aware of->

These two classes are now capable of displaying and handling arrow-based
decompositions.

* ENH: Increase arrow type awareness

This changeset makes a number of view controllers aware of the
decompositions of arrow type. In addition it makes DecompositionView
better support this variant.

It also deals with the cases where the controllers shouldn't modify the
views.

* BUG: Fix raycasting for multiple decompositions

This patch adds the ability to check for raycasting over multiple
decompositions. In addition it subclasses the THREE.ArrowHelper class to
implement the raycast method so that collisions can be detected and
attributed.

* BUG: Include missing assignment

* ENH: Improve the API of the controller

* BUG:Don't use _.pick, it's not supported in the nb

* ENH: Make Emperor aware of biplots

This commit adds a fair number of changes, for the most part the intent
is to be able to interpret and process feature data as transported by
the ordination. The constructor can now take as a parameter feature
metadata. As a result when serializing the object into a JSON object, we
now use a single object, as returned by _to_dict. The main template was
also modified to account for this change. _to_dict puts the sample and
feature ordinations in a common format for easy consumption by the
JavaScript API.

A number of methods have been modified to become aware about feature
metadata and the feature ordination. Specifically:

- _validate_metadata: takes as a parameter the metadata it is validating
and allows for a case where no metadata is provided.

- _process_data: checks if there's a feature ordination and if there is
one proceeds to create the appropriate objects and convert them into
JSON-serializeable types.

- _to_legacy_map now takes as a parameter the metadata to convert.

* STY: Fix stylistic issues with JS

* TST: Fix tests to update the template

* TST: Add tests for the constructor & biplots

* TST: Add tests for _process_data

* ENH: Remove unused method

* TST: Add simple group of tests for most changes

* TST/BUG: Add more tests and fix bug in arrows

The DecompositionView object had a bug where the length of the arrows
was calculated based on the length of the normalized vector not of the
original vector.

Additionally this patch adds some tests to check for the sanity of the
DecompositionView object when initialized with an ArrowType.

* DOC: Add example notebook for biplots

* STY: Fix stylistic issues with JavaScript

* ENH: Add the ability to process procrustes plots

* ENH:Make the GUI objects aware of procrustes plots

Adds a couple of methods for the model and the views to process an edge
list that's used to display procrustes plots.

This patch also includes some cleanup to the repeated code in the
changeVisibleDimensions and flipDimension methods. See updatePositions.

Draw.js has a function to draw lines as a THREE.LineSegments object,
this is recommended from their documentation to be more efficient at
rendering, etc. This used to be a problem when we had a large number of
samples being compared at the same time.

* DOC: Add example notebook for procrustes

* TST: Add tests for procrustes

* TST: Write tests for javascript portions

Fix documentation

* ENH: Add the ability to color procrustes edges

This has also created a number of methods in the AxesController class to
deal with color changes in different elements.

* ENH: Cleanup test suite

* BUG: Fix problem with settings

* BUG: Bring back to 5

* BUG: Should be 7 instead of 5

* ENH: Improve build.py to include subjects

* ENH/BUG: Add label for arrows

* Address review comments by @wasade

* TST: Add polyfill for padStart

* TST/BUG: Add additional JavaScript polyfills

* STY: Fix stylistic errors in JavaScript

* ENH: Add example of how to calculate biplots

* DOC: Document the useful categories in the 📓

* ENH: Fix HTML for axes and .ready for controller

The HTML code for the procrustes controls was missing a few closing
tags. I've also changed the formatting so it is easier to read.

The _expected and _seen attributes in the main controller are now
updated through the addTab method, so there's no need to update these
values manually.

Addressing @antgonza's review comments