Fixes problem with the camera not being centered (#706)
* WIP: Improve camera positioning

The camera was currently being positioned at the origin, instead of at
the center of the current data on screen.

* WIP: I have not been able to reset the camera

* BUG: Remove control damping

When damping is enabled it makes the recenternig the orbit controls go
to a point close to the center but not exactly it. This is because the
object is trying to simulate inertia during the recentering. I have now
disabled it.

* TST: Expand unit tests for ScenePlotView3D

* STY: Fix stylistic issues