Controllers set applicable parameters on switching back and forth between scatter and parallel-plot
EmperorViewControllerABC adds a forceRefresh function, and defaults it to a toJSON followed by a fromJSON.
ScenePlotView3D's checkUpdate now returns a bit vector of required actions rather than a single boolean.
This is handled in new controller logic.
Attribute setters updated to set attributes for with correct indexing in parallel plot mode.
Shape changes are ignored
Axes changes are ignored