TST: Remove "scripts/*.py" from flake8 travis call
...since there is no scripts/ folder in Emperor any more.

This is really bizarre, though -- this was definitely not failing
last week: see
https://travis-ci.com/github/biocore/emperor/jobs/339150405.

It looks like the flake8 version was updated between python 3.6 builds
on Travis (3.7.9 to 3.8.2), but I have v3.7.9 installed on my computer
and running the flake8 command here still gives me an error about
scripts/ not being a real directory.

Uh, I think the path of least resistance is just fixing this for right
now, but I'm very confused.